+++ src/arch/arm64/memory/address_space.rs
@@ -16,14 +16,14 @@ use libkernel::{
             L0Table, MapAttributes, MappingContext, map_range,
         },
         pg_tear_down::tear_down_address_space,
-        pg_walk::{WalkContext, get_pte, walk_and_modify_region},
+        pg_walk::{get_pte, walk_and_modify_region},
     },
     error::{KernelError, MapError, Result},
     memory::{
         PAGE_SIZE,
         address::{TPA, VA},
         page::PageFrame,
-        paging::{PaMapper, PageAllocator, PageTableEntry, PgTableArray, permissions::PtePermissions},
+        paging::{PaMapper, PageAllocator, PageTableEntry, PgTableArray, permissions::PtePermissions, walk::WalkContext},
         proc_vm::address_space::{PageInfo, UserAddressSpace},
         region::{PhysMemoryRegion, VirtMemoryRegion},
     },
